@@ -21,42 +21,45 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^6.0.3",
-		"@angular/common": "^6.0.3",
-		"@angular/compiler": "^6.0.3",
+		"@angular/animations": "^8.2.14",
+		"@angular/common": "^8.2.14",
+		"@angular/compiler": "^8.2.14",
 		"@angular/core": "^8.2.14",
-		"@angular/forms": "^6.0.3",
-		"@angular/http": "^6.0.3",
-		"@angular/platform-browser": "^6.0.3",
-		"@angular/platform-browser-dynamic": "^6.0.3",
-		"@angular/router": "^6.0.3",
+		"@angular/forms": "^8.2.14",
+		"@angular/http": "^7.2.15",
+		"@angular/platform-browser": "^8.2.14",
+		"@angular/platform-browser-dynamic": "^8.2.14",
+		"@angular/router": "^8.2.14",
 		"chart.js": "2.9.3",
-		"core-js": "^2.5.4",
+		"core-js": "3.6.1",
 		"lodash": "4.17.15",
 		"quagga": "0.12.1",
 		"rxjs": "^6.0.0",
-		"zone.js": "^0.8.26"
+		"zone.js": "~0.9.1"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "~0.6.8",
-		"@angular-devkit/build-ng-packagr": "~0.6.8",
-		"@angular/cli": "^8.3.21",
-		"@angular/compiler-cli": "^6.0.3",
+		"@angular-devkit/build-angular": "^0.803.21",
+		"@angular-devkit/build-ng-packagr": "^0.803.21",
+		"@angular/cli": "8.3.21",
+		"@angular/compiler-cli": "^8.2.14",
 		"@angular/language-service": "^6.0.3",
 		"@types/jasmine": "~2.8.6",
 		"@types/jasminewd2": "~2.0.3",
+		"@types/lodash": "^4.14.149",
 		"@types/node": "~8.9.4",
 		"jasmine-core": "~2.99.1",
 		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "~1.7.1",
+		"karma": "^4.4.1",
 		"karma-chrome-launcher": "~2.2.0",
 		"karma-coverage-istanbul-reporter": "~2.0.0",
 		"karma-jasmine": "~1.1.1",
 		"karma-jasmine-html-reporter": "^0.2.2",
-		"protractor": "~5.3.0",
+		"ng-packagr": "^4.0.0",
+		"protractor": "^6.0.0",
 		"ts-node": "~5.0.1",
+		"tsickle": "^0.34.1",
 		"tslib": "^1.7.1",
 		"tslint": "~5.9.1",
-		"typescript": "~2.7.2"
+		"typescript": "~3.5.0"
 	}
 }
