@@ -3,7 +3,8 @@ import { NgModule, APP_INITIALIZER } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { NvmSettingsModule, NvmSettingsService } from 'nvm-settings';
-import { NvmChartsModule } from 'projects/nvm-charts/src/public_api';
+import { AppRoutingModule } from './app-routing.module';
+import { RouterModule } from '@angular/router';
 
 export const settingsProvider = (config: NvmSettingsService) => () => {
 	return config.load('/assets/settings.json');
@@ -12,13 +13,20 @@ export const settingsProvider = (config: NvmSettingsService) => () => {
 export const useAppConfigProvider = { provide: APP_INITIALIZER, useFactory: settingsProvider, deps: [NvmSettingsService], multi: true };
 
 @NgModule({
-	declarations: [AppComponent],
+	declarations: [
+		AppComponent
+	],
 	imports: [
+		RouterModule,
 		BrowserModule,
-		NvmChartsModule,
+		AppRoutingModule,
 		NvmSettingsModule.forRoot()
 	],
-	providers: [useAppConfigProvider],
-	bootstrap: [AppComponent]
+	providers: [
+		useAppConfigProvider
+	],
+	bootstrap: [
+		AppComponent
+	]
 })
-export class AppModule {}
+export class AppModule { }
